Refactor: move interfaces into separate folder
