+++ solutions/musa-yuksel/src/interfaces/IScoreboard.ts
@@ -1,0 +1,7 @@
+import { IPlayer } from './IPlayer';
+
+export interface IScoreboard extends IPlayer {
+  isDeuce: boolean;
+  hasWon: boolean;
+  hasAdvantage: boolean;
+}
