+++ solutions/musa-yuksel/src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export type { IPlayer } from './IPlayer';
+export type { IScoreboard } from './IScoreboard';
